create a stored procedure AddBonus that adds a new correction for a student

+++ 0x00-MySQL_Advanced/6-bonus.sql
@@ -1,7 +1,7 @@
 -- creates a stored procedure AddBonus that adds a new correction for a student
 
 DELIMITER //
-CREATE PROCEDURE AddBonus(IN user_id INT, IN project_name VARCHAR(255), IN score INT)
+CREATE PROCEDURE AddBonus(IN user_id INT, IN project_name VARCHAR(255), IN score FLOAT)
 BEGIN 
 	DECLARE @project_id INT;
 	DECLARE @count INT;
